LayoutTests/imported/w3c/web-platform-tests/html/semantics/the-button-element/command-and-commandfor/button-event-dispatch-expected.txt



FAIL event dispatches on click with addEventListener assert_equals: composed expected false but got true
FAIL event dispatches on click with oncommand property assert_equals: composed expected false but got true
FAIL setting custom command property to --foo (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --foo (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --foo- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --foo- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --cAsE-cArRiEs (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --cAsE-cArRiEs (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to -- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to -- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --a- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --a- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --a-b (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --a-b (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --- (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command property to --show-picker (must include dash) sets event command assert_equals: composed expected false but got true
FAIL setting custom command attribute to --show-picker (must include dash) sets event command assert_equals: composed expected false but got true
PASS setting custom command property to -foo (no dash) did not dispatch an event
PASS setting custom command attribute to -foo (no dash) did not dispatch an event
PASS setting custom command property to -foo- (no dash) did not dispatch an event
PASS setting custom command attribute to -foo- (no dash) did not dispatch an event
PASS setting custom command property to foo-bar (no dash) did not dispatch an event
PASS setting custom command attribute to foo-bar (no dash) did not dispatch an event
PASS setting custom command property to -foo bar (no dash) did not dispatch an event
PASS setting custom command attribute to -foo bar (no dash) did not dispatch an event
PASS setting custom command property to —-emdash (no dash) did not dispatch an event
PASS setting custom command attribute to —-emdash (no dash) did not dispatch an event
PASS setting custom command property to hidedocument (no dash) did not dispatch an event
PASS setting custom command attribute to hidedocument (no dash) did not dispatch an event
PASS event does not dispatch if click:preventDefault is called
PASS event does not dispatch on input[type=button]
PASS event does not dispatch if invoker is disabled
PASS event does NOT dispatch if button is form associated, with implicit type
PASS event does NOT dispatch if button is form associated, with explicit type=invalid
FAIL event dispatches if button is form associated, with explicit type=button assert_equals: composed expected false but got true
PASS event does NOT dispatch if button is form associated, with explicit type=submit
PASS event does NOT dispatch if button is form associated, with explicit type=reset
PASS event dispatches if invokee is non-HTML Element
